[feat] : 게시글 상세 작성자 추가, 게시글 삭제 예외처리 추가

+++ src/main/java/com/back/domain/post/service/PostService.java
@@ -18,6 +18,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -34,7 +35,7 @@ public class PostService {
 
         // 카테고리 변환 예외 처리
         Post.Category category = Post.Category.from(dto.category())
-                .orElseThrow(() -> new ServiceException("BAD_REQUEST", "유효하지 않은 카테고리입니다."));
+                .orElseThrow(() -> new ServiceException("400", "유효하지 않은 카테고리입니다."));
 
         Post post = Post.builder()
                 .title(dto.title())
@@ -53,16 +54,20 @@ public class PostService {
     @Transactional
     public RsData<String> deletePost(Long postId) {
         Member member = getCurrentMemberOrThrow();
-        Post post = getPostOrThrow(postId);
 
+        Optional<Post> optionalPost = postRepository.findById(postId);
+        if (optionalPost.isEmpty()) {
+            throw new ServiceException("404", "이미 삭제되었거나 존재하지 않는 게시글입니다.");
+        }
+
+        Post post = getPostOrThrow(postId);
         // 본인 게시글인지 확인
         if (!post.getMember().getId().equals(member.getId())) {
-            throw new ServiceException("FORBIDDEN", "자신의 게시글만 삭제할 수 있습니다.");
+            throw new ServiceException("403", "자신의 게시글만 삭제할 수 있습니다.");
         }
 
         postRepository.delete(post);
-
-        return new RsData<>("SUCCESS", "게시글 삭제 완료", null);
+        return new RsData<>(ResultCode.SUCCESS, "게시글 삭제 완료", null);
     }
 
 
@@ -131,8 +136,6 @@ public class PostService {
         }
     }
 
-
-
     //찜 목록 조회
     @Transactional(readOnly = true)
     public List<PostListDTO> getFavoritePosts() {
@@ -150,7 +153,7 @@ public class PostService {
     private Member getCurrentMemberOrThrow() {
         Member member = rq.getMember();
         if (member == null) {
-            throw new ServiceException("UNAUTHORIZED", "로그인이 필요합니다.");
+            throw new ServiceException("401", "로그인이 필요합니다.");
         }
         return member;
     }
@@ -158,13 +161,13 @@ public class PostService {
     // 게시글 조회 에러
     private Post getPostOrThrow(Long postId) {
         return postRepository.findById(postId)
-                .orElseThrow(() -> new ServiceException("NOT_FOUND", "게시글이 존재하지 않습니다."));
+                .orElseThrow(() -> new ServiceException("404", "게시글이 존재하지 않습니다."));
     }
 
     // 찜 기능 시 동기화 문제 처리 락
     private Post getPostForUpdateOrThrow(Long postId) {
         return postRepository.findByIdForUpdate(postId)
-                .orElseThrow(() -> new ServiceException("NOT_FOUND", "오류 입니다."));
+                .orElseThrow(() -> new ServiceException("404", "오류 입니다."));
     }
 
 }
